Task/requirements_complete.txt
Flask>=2.0.0
mysql-connector-python>=8.0.0
cryptography>=3.4.0
joblib>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
python-dateutil>=2.8.0

